requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.1.0
beautifulsoup4==4.13.3
braceexpand==0.1.7
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
docker-pycreds==0.4.0
einops==0.8.1
emoji==2.14.1
executing==2.2.0
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.2
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
hf-xet==1.1.2
huggingface-hub==0.32.4
idna==3.10
inquirerpy==0.3.4
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
lightning==2.5.0.post0
lightning-utilities==0.12.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pfzy==0.3.4
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pynvml==12.0.0
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
sentry-sdk==2.21.0
setproctitle==1.3.4
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
soupsieve==2.6
stack-data==0.6.3
stanza==1.10.1
sympy==1.13.1
tensorboardX==2.6.2.2
threadpoolctl==3.6.0
tokenizers==0.21.0
torch==2.6.0
torchmetrics==1.6.1
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.1
triton==3.2.0
typing_extensions==4.13.1
tzdata==2025.1
urllib3==2.3.0
wandb==0.19.6
wcwidth==0.2.13
webdataset==0.2.111
wheel==0.45.1
yarl==1.18.3
